% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cell.occur}
\alias{cell.occur}
\title{Plot of cell occurrences}
\usage{
cell.occur(nums)
}
\arguments{
\item{nums}{Results of \code{deconvoluting}.}
}
\value{
This function will return a plot showing the occurrence of cells as well as a global variable named "occurred.cells"
showing the names of the cells occurred.
}
\description{
Plot of occurrences of each cell(cell type) in the whole spatial transcriptome.
}
