% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{cell.sampling}
\alias{cell.sampling}
\title{Sample single cells.}
\usage{
cell.sampling(ncells, annotations = NULL, size, prot = T)
}
\arguments{
\item{ncells}{Number of cells in raw scRNA-seq data.}

\item{annotations}{A \code{data.frame} with 2 columns: the first column should be barcodes,
the second column should be the corresponding cell type. If missing, complete random sampling will be run.}

\item{size}{Total number of cells you want to take out.}

\item{prot}{Protection of cell types with low occurrence in raw data. If \code{prot = T},
it will be ensured that there is at least one cell of each cell type in the sampling results;
otherwise due to \code{size}, some cell types may not appear. Default is \code{T}.}
}
\value{
If annotations are provided, this function will return a \code{data.frame} containing
barcodes of selected cells and their annotations, and its \code{rownames} are the original indices.
Otherwise only indices.
}
\description{
Sample some single cells from original scRNA-seq data. If cell annotations are provided,
this function will run stratified sampling; otherwise it will run random sampling.
}
