% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cell.type.weight}
\alias{cell.type.weight}
\title{#' Heatmap of any correlation
#'
#' This function
#'
#' @param x The data whose correlation will be calculated and visualized. One column represents one variable.
#' @param group
#' @param method Method of calculating correlation coefficient.
#'
#' @return A global variable named "corr", represents for correlation of cell type
#' proportion across spatial locations and its visualization.
#'
#' @export
corr.heatmap = function(x, group, method = c("pearson","spearman")){
x = as.matrix(x)
corr <<- Hmisc::rcorr(x,type=method)[\link{"r"}]
corrtp = reshape2::melt(corr)
ggplot(corrtp,aes(x=Var1,y=Var2,fill=value))+
geom_tile()+
scale_fill_gradient2(low="White", high="Blue")+
labs(x=NULL,y=NULL,title = "correlation")+
theme_bw(base_size = 15)+
theme(axis.text.x = element_blank(),
axis.text.y = element_blank())
}
Plot of weights for each cell type}
\usage{
cell.type.weight(nums, cell.type, annotations, cell.names = NULL, coords, name)
}
\arguments{
\item{nums}{Results of \code{deconvoluting}.}

\item{cell.type}{Whether to merge \code{nums} by cell type.}

\item{annotations}{If \code{cell.type = F}, this should be the annotation of the single cells.}

\item{cell.names}{Name of cells (or cell types) to be plotted. Default is all.}

\item{coords}{Coordinates of spatial spots. \code{rownames} should be the barcode(name) of each spot
and \code{colnames} should be "x" and "y".}

\item{name}{The name of this pdf file.}
}
\description{
Plot of weights for each cell type
}
