% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{coloc.corr}
\alias{coloc.corr}
\title{Colocalization of cell types}
\usage{
coloc.corr(nums, cell.type, annotations, method = c("pearson", "spearman"))
}
\arguments{
\item{nums}{Results of \code{deconvoluting}.}

\item{cell.type}{Whether to merge \code{nums} by cell type. If \code{TRUE}, a \code{cell type * cell type}
correlation matrix will be returned.}

\item{annotations}{If \code{cell.type = F}, this should be the annotation of the single cells.}

\item{method}{Method of calculating correlation coefficient.}
}
\value{
A global variable named "corr", represents for correlation of cell type
proportion across spatial locations and its visualization.
}
\description{
This function plots the correlation matrix of cell abundance, to infer colocaliaztion of cell types.
}
