% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{coloc.network}
\alias{coloc.network}
\title{Colocalization network}
\usage{
coloc.network(corr, thresh, cell.type, annotations, ntypes, color = NULL)
}
\arguments{
\item{corr}{Correlation matrix}

\item{thresh}{Threshold of correlation coefficient. Value large than this value will be regarded as an edge.}

\item{cell.type}{Whether the result is by cell-type (\code{TRUE}) or by single-cell (\code{FALSE}).}

\item{annotations}{If \code{cell.type = T}, this should be the names of the cell types.
If \code{cell.type = F}, this should be the annotation of the single cells.}

\item{ntypes}{Number of cell types.}

\item{color}{The colors of cell types. Default is \code{rainbow(ntypes)}.}
}
\value{
An "igraph" object
}
\description{
This function gains an igraph network to show colocalization of cells (cell types).
}
