% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{deconvoluting}
\alias{deconvoluting}
\title{Processing deconvolution}
\usage{
deconvoluting(
  ref,
  st,
  cellnames = NULL,
  genemode,
  gene.list,
  var_thresh = 0.025,
  exp_thresh = 0.03,
  hpmode,
  hp,
  normalize = T,
  thre = 1e-10,
  dopar = T,
  ncores,
  realtime = F,
  dir = NULL
)
}
\arguments{
\item{ref}{Reference, the scRNA-seq data. Pseudo-count will be added and counts will be normalized to TPM.}

\item{st}{Spatial transcriptomics to be deconvoluted}

\item{cellnames}{(optional) Cells(Cell types) to be used for deconvolution in reference.
If missing, all cells will be used.}

\item{genemode}{Mode of genes to be included in deconvolution. Three modes are available: "\code{default}", using the intersect of \code{ref} and \code{st};
"\code{customized}", indicating the genes yourself; "\code{filtered}", use some parameters to filter the genes (see \code{var_thresh} and \code{exp_thresh}
for details).}

\item{gene.list}{If \code{genemode = "customized"}, this will be the genes you want to use in deconvolution.}

\item{var_thresh}{If \code{genemode = "filtered"}, this will be the threshold of variance of gene expression in reference.}

\item{exp_thresh}{If \code{genemode = "filtered"}, this will be the lowest average gene expression per spot in spatial transcriptomics.}

\item{hpmode}{Mode of choosing hyperparameter adjusting the proportion of
LS term and penalty term. Three modes are available: "\code{default}",  using the default
hyperparameter we set according to the number of cells and genes; "\code{customized}",
setting the hyperparameter yourself; "\code{autoselection}", automatically calculating
and selecting the optimal hyperparameter(may take a long time).}

\item{hp}{If \code{hpmode = "customized"}, this will be the hyperparameter you want to use.}

\item{normalize}{Whether to normalize the reference or not. \code{TRUE} is the default value and is recommended when deconvoluting spatial transcriptomics.}

\item{thre}{Threshold, numbers less than this value in the result will be regarded as zero.
Default is \code{1e-10}.}

\item{dopar}{Whether to use parallel computing. \code{TRUE} is recommended when
\code{hpmode = "autoselection"}.}

\item{ncores}{Number of cores to be used when doing parallel computing.}

\item{realtime}{Whether to write results into disk in real time.}
}
\value{
A matrix, showing the number of estimated absolute abundance of each cell(cell type) in each spot.
}
\description{
Processing deconvolution, getting the number of each reference cell(cell type)
in each spot of spatial transcriptomics.
}
