% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{extract.seurat.st}
\alias{extract.seurat.st}
\title{Seurat interface 2}
\usage{
extract.seurat.st(st_seu, ...)
}
\arguments{
\item{st_seu}{Seurat object of spatial transcriptomics}

\item{...}{Additional parameters for \code{Seurat::GetTissueCoordinates}}
}
\value{
A list, containing single-cell expression matrix and annotations.
}
\description{
Seurat interface for spatial transcriptomics. This function will extract the expression matrix and spatial coordinates from the Seurat object.
}
\seealso{
\code{\link[=extract.seurat.sc]{extract.seurat.sc()}}
}
