% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{extract.subgraph}
\alias{extract.subgraph}
\title{Extracting subgraph}
\usage{
extract.subgraph(g, vertex_of_interest, ...)
}
\arguments{
\item{g}{The "igraph" object, result of \code{coloc.network}.}

\item{...}{Other parameters passed to \code{plot}.}

\item{node}{The node to be chosen. The name of that cell or cell type.}
}
\value{
An "igraph" object of the subgraph. Also, this function will plot the subgraph with the chosen node in the center.
}
\description{
This function extracts a subgraph, including a chosen node and its all neighbors, from the co-localization network generated by \code{coloc.network}.
}
