% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{gene.filter}
\alias{gene.filter}
\title{Getting differentially expressed genes}
\usage{
gene.filter(ref, st, var_thresh = 0.025, exp_thresh = 0.003)
}
\arguments{
\item{ref}{Reference.}

\item{st}{Spatial transcriptomics data.}

\item{var_thresh}{The threshold of variance of gene expression in reference.}

\item{exp_thresh}{Lowest average gene expression per spot in spatial transcriptomics.}
}
\description{
Getting differentially expressed genes across cells(cell types).
The expression of DE genes shouldn't be too low in both scRNA-seq data
and spatial transcriptomics data, and should have some difference across cells(cell types).
}
