% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{get.ref}
\alias{get.ref}
\title{Getting the gene expression profile of each cell type.}
\usage{
get.ref(sc, annotations, gene.list = NULL, dopar = T, ncores)
}
\arguments{
\item{sc}{Raw scRNA-seq data. Note that colnames of sc must be cell barcodes.}

\item{annotations}{A \code{data.frame} with 2 columns: the first column should be barcodes,
the second column should be the corresponding cell type.}

\item{gene.list}{(optional) Genes to be calculated. If missing, all genes
will be calculated.}
}
\value{
A matrix showing the gene expression profile of each cell type,
one row represents one cell type and one column represents one gene.
}
\description{
If you want to use the gene expression profile of each cell type as reference,
this function can calculate the average expression of genes within all cells
of every cell type.
}
