% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{hNpp}
\alias{hNpp}
\title{Calculating Entropy and Perplexity}
\usage{
hNpp(nums, thre = 1e-10)
}
\arguments{
\item{nums}{Results.}
}
\value{
A matrix with 2 rows: the first is Entropy, the second is Perplexity.
}
\description{
Calculating Entropy and Perplexity for each spot.
}
