% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{profile.imputation}
\alias{profile.imputation}
\title{Spatial expression profile imputation}
\usage{
\method{profile}{imputation}(abnum, sc)
}
\arguments{
\item{abnum}{The result of \code{to.absolute.abundance} !}

\item{sc}{Single-cell expression profile.}
}
\value{
A matrix with dimensions same as \code{st}, the reconstructed expression profile.
}
\description{
Reconstruct spatial expression profile by estimated (abosulte) cell abundance and single-cell expression profile.
}
