% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{profile_comparison}
\alias{profile_comparison}
\title{Corrected profile v.s. original profile}
\usage{
profile_comparison(ref, ests, barcodes = NULL)
}
\arguments{
\item{ref}{Original reference.}

\item{ests}{Corrected expression profile.}

\item{barcodes}{(Optional) the barcodes of the cells whose profile is to compared. Default is all.}
}
\value{
A pdf file named "profile_comparison".
}
\description{
Comparison of the corrected and original profile.
}
