% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Redeconve.R
\docType{package}
\name{Redeconve}
\alias{Redeconve-package}
\alias{Redeconve}
\title{Redeconve: Deconvolution of spatial transcriptome with single cell resolution.}
\description{
Deconvolution of spatial transcriptome with single cell resolution.
}
\author{
\strong{Maintainer}: Zixiang Zhou \email{zxclbnzhou@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
