% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{resids}
\alias{resids}
\title{Calculating residuals}
\usage{
resids(nums, ref, st, mode, genes, inside = "out")
}
\arguments{
\item{nums}{Result generated by \code{deconvoluting}.}

\item{ref}{Reference.}

\item{st}{Spatial transcriptomics data.}

\item{mode}{Mode of calculating residuals. "\code{spot}" for spot-wise and "\code{gene}" for gene-wise.}

\item{genes}{(optional) If \code{mode = "gene"}, this will be a gene or a list of genes whose
residuals you want to calculate. The default is all genes.}

\item{inside}{Indicates whether normalization is required. You don't need to adjust it when using.}
}
\description{
Calculating residuals for the whole spatial transciptome or some genes.
}
