% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{sc2type}
\alias{sc2type}
\title{Converting single cell results to cell type results}
\usage{
sc2type(nums, annotations)
}
\arguments{
\item{nums}{The results of single-cell resolution}

\item{annotations}{The annotation of the single cells.}
}
\description{
This function converts single-cell abundance to cell-type abundance.
}
