% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{show.cellsofinterest}
\alias{show.cellsofinterest}
\title{Showing cells of interest}
\usage{
show.cellsofinterest(nums, occurred.cells)
}
\arguments{
\item{nums}{Estimated cell numbers.}

\item{occurred.cells}{Barcodes of cells whose abundance is not zero. This value can be generated by \code{cell.occur}.}
}
\value{
a plot showing the mean abundance vs. coefficient of variation (standard deviation divided by mean value) of all occurred cells. Cells with high mean abundance and/or high variance will be labeled.
}
\description{
Summary of the result, for users to identify cells of interest.
}
