% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{spatial.cell.number}
\alias{spatial.cell.number}
\title{theme_bw(base_size = 15)+
Plot of numbers for each given cell/cell type}
\usage{
spatial.cell.number(nums, cell.names = NULL, coords, size = 1, name)
}
\arguments{
\item{nums}{Estimated cell numbers.}

\item{cell.names}{Name of cells (or cell types) to be plotted. Default is all.}

\item{coords}{Coordinates of spatial spots. \code{rownames} should be the barcode(name) of each spot
and \code{colnames} should be "x" and "y".}

\item{size}{Size of spot. Default is 1.}

\item{name}{The name of this pdf file.}
}
\description{
Plot of numbers for each cell cell/cell type.
}
