% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{spatial.piechart}
\alias{spatial.piechart}
\title{This is the CARD function}
\usage{
spatial.piechart(
  nums,
  coords,
  colors = NULL,
  img.backgound = F,
  img,
  title = NULL
)
}
\description{
This is the CARD function
}
