% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{to.proportion}
\alias{to.proportion}
\title{Giving results interpretability}
\usage{
to.proportion(res)
}
\arguments{
\item{res}{Result of \code{deconvoluting}.}
}
\value{
A cell-by-spot matrix, with the sum of each column equaling to 1.
}
\description{
One way of gaining interpretability, normalize results to proportion
}
