% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{cell.type.weight}
\alias{cell.type.weight}
\title{Plot of weights for each cell type}
\usage{
cell.type.weight(nums, cell.type, annotations, cell.names = NULL, coords, name)
}
\arguments{
\item{nums}{Results of \code{deconvoluting}.}

\item{cell.type}{Whether to merge \code{nums} by cell type.}

\item{annotations}{If \code{cell.type = F}, this should be the annotation of the single cells.}

\item{cell.names}{Name of cells (or cell types) to be plotted. Default is all.}

\item{coords}{Coordinates of spatial spots. \code{rownames} should be the barcode(name) of each spot
and \code{colnames} should be "x" and "y".}

\item{name}{The name of this pdf file.}
}
\description{
Plot of weights for each cell type
}
