% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{coloc.network}
\alias{coloc.network}
\title{Visualization of colocalization}
\usage{
coloc.network(corr, thresh, cell.type, annotations, ntypes, color = NULL, name)
}
\arguments{
\item{corr}{Correlation matrix returned by \code{coloc.corr}}

\item{thresh}{Threshold of correlation coefficient. Value large than this value will be regarded as an edge.}

\item{cell.type}{Whether the result is by cell-type (\code{TRUE}) or by single-cell (\code{FALSE}).}

\item{annotations}{If \code{cell.type = T}, this should be the names of the cell types.
If \code{cell.type = F}, this should be the annotation of the single cells.}

\item{ntypes}{Number of cell types.}

\item{color}{The colors of cell types. Default is \code{rainbow(ntypes)}.}

\item{name}{The name of this pdf file.}
}
\description{
Draws a network to show colocalization of cells (cell types).
}
