% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{extract.seurat.sc}
\alias{extract.seurat.sc}
\title{Seurat interface 1}
\usage{
extract.seurat.sc(sc_seu)
}
\arguments{
\item{sc_seu}{Seurat object of single-cell transcriptomics}
}
\value{
A list, containing single-cell expression matrix and annotations.
}
\description{
Seurat interface for single-cell transcriptomics. This function will extract the expression matrix and annotations (as "Idents") from the Seurat object.
}
\seealso{
\code{\link[=extract.seurat.st]{extract.seurat.st()}}
}
