% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{extract.seurat.sc}
\alias{extract.seurat.sc}
\title{Seurat interface 1}
\usage{
extract.seurat.sc(sc_seu, anno_slot)
}
\arguments{
\item{sc_seu}{Seurat object of single-cell transcriptomics}

\item{anno_slot}{Character, name of slot storing annotations information, under "meta.data" slot.}
}
\value{
A list, containing single-cell expression matrix and annotations.
}
\description{
Seurat interface for single-cell transcriptomics
}
