% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.R
\name{extract.seurat.st}
\alias{extract.seurat.st}
\title{Seurat interface 2}
\usage{
extract.seurat.st(st_seu, coord_slot = "image")
}
\arguments{
\item{coord_slot}{Character, name of slot storing annotations information, under "image" slot. Default is "image".}

\item{sc_seu}{Seurat object of spatial transcriptomics}
}
\value{
A list, containing single-cell expression matrix and annotations.
}
\description{
Seurat interface for spatial transcriptomics
}
