% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{show.cellsofinterest}
\alias{show.cellsofinterest}
\title{Showing cells of interest}
\usage{
show.cellsofinterest(nums, occurred.cells, outliernum = 10)
}
\arguments{
\item{nums}{Estimated cell numbers.}

\item{occurred.cells}{Barcodes of cells whose abundance is not zero. This value can be generated by \code{cell.occur}.}

\item{outliernum}{How many outliers to display. Default is 10.}
}
\value{
a plot showing the mean abundance vs. standard deviation of all occurred cells. Cells with high mean abundance and/or high variance will be labeled.
}
\description{
For users to identify cells of interest.
}
