% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{spatial.cell.number}
\alias{spatial.cell.number}
\title{Plot of numbers for each given cell/cell type}
\usage{
spatial.cell.number(
  nums,
  cell.names = NULL,
  coords,
  size = 1,
  img.background = F,
  img,
  pdfout = T
)
}
\arguments{
\item{nums}{Estimated cell numbers.}

\item{cell.names}{Name of cells (or cell types) to be plotted. Default is all.}

\item{coords}{Coordinates of spatial spots. \code{rownames} should be the barcode(name) of each spot
and \code{colnames} should be "x" and "y".}

\item{size}{Size of spot. Default is 1.}

\item{img.background}{Whether to use an image as background. Default is \code{FALSE}.}

\item{img}{Image to be used as background.}

\item{pdfout}{Whether to draw the plot in a pdf file named "spatial_cell_number". When more than one cell is plotted, this is recommended to be \code{TRUE}.}
}
\value{
A list containing the spatial distribution of each given cell.
}
\description{
Plot of numbers for each given cell/cell type.
}
