% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{spatial.gene}
\alias{spatial.gene}
\title{Spatial expression}
\usage{
spatial.gene(st, coords, gene.list)
}
\arguments{
\item{st}{Spatial transcriptomics.}

\item{coords}{Coordinates of spatial spots.}

\item{gene.list}{List of genes.}
}
\value{
A pdf file named "spatial_genes".
}
\description{
Visualization of spatial expression of some genes.
}
