% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{spot.pie}
\alias{spot.pie}
\title{Pie chart of a single spot}
\usage{
spot.pie(num, colors = NULL, title)
}
\arguments{
\item{num}{One column of the results.}

\item{colors}{The colors of cell types. Default is \code{rainbow(ntypes)}.}

\item{title}{Title of this plot.}
}
\description{
Pie chart of a single spot.
}
